unlockDialog: Show media notifications on lock screen
While we have included a built-in media section in the calendar for
a while, it is currently only available when the session is unlocked.
The ability to quickly pause or resume media playback directly from
the lock screen is quite handy though, so start showing them alongside
the existing screen shield notifications.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/488
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3306>